Renaming of return and notification processing
