+++ wirecardpaymentgateway/classes/Engine/ReturnResponse.php
@@ -33,36 +33,27 @@
  * @license GPLv3
  */
 
-namespace WirecardEE\Prestashop\Classes\ResponseProcessing;
+namespace WirecardEE\Prestashop\Classes\Engine;
 
-use Wirecard\PaymentSdk\Response\InteractionResponse;
+use Wirecard\PaymentSdk\Response\Response;
 
 /**
- * Class InteractionResponseProcessing
- * @package WirecardEE\Prestashop\Classes\ResponseProcessing
+ * Class ReturnResponse
+ *
+ * @package WirecardEE\Prestashop\Classes\Engine
  * @since 2.1.0
  */
-final class InteractionResponseProcessing implements ResponseProcessing
+final class ReturnResponse extends PaymentSdkResponse
 {
-    /** @var InteractionResponse  */
-    private $response;
-
     /**
-     * InteractionResponseProcessing constructor.
-     *
-     * @param InteractionResponse $response
+     * @param array $response
+     * @return Response|false
      * @since 2.1.0
      */
-    public function __construct($response)
+    public function process($response)
     {
-        $this->response = $response;
-    }
+        parent::process($response);
 
-    /**
-     * @since 2.1.0
-     */
-    public function process()
-    {
-        \Tools::redirect($this->response->getRedirectUrl());
+        return $this->backend_service->handleResponse($response);
     }
 }
