+++ wirecardpaymentgateway/classes/Response/Cancel.php
@@ -33,19 +33,54 @@
  * @license GPLv3
  */
 
-namespace WirecardEE\Prestashop\Classes\EngineResponseProcessing;
+namespace WirecardEE\Prestashop\Classes\Response;
+
+use WirecardEE\Prestashop\Helper\Service\ContextService;
+use WirecardEE\Prestashop\Helper\Service\OrderService;
+use WirecardEE\Prestashop\Helper\OrderManager;
 
 /**
- * Interface EngineResponseProcessing
- *
- * @package WirecardEE\Prestashop\Classes\EngineResponseProcessing
+ * Class Cancel
  * @since 2.1.0
+ *@package WirecardEE\Prestashop\Classes\Response
  */
-interface EngineResponseProcessing
+final class Cancel implements ProcessablePaymentResponse
 {
+    /** @var \Order */
+    private $order;
+
+    /** @var ContextService */
+    private $context_service;
+
+    /** @var OrderService */
+    private $order_service;
+
     /**
-     * @param array|string $response
+     * CancelResponseProcessing constructor.
+     *
+     * @param \Order $order
+     */
+    public function __construct($order)
+    {
+        $this->order = $order;
+        $this->context_service = new ContextService(\Context::getContext());
+        $this->order_service = new OrderService($order);
+    }
+
+    /**
+     * @throws \Exception
      * @since 2.1.0
      */
-    public function process($response);
+    public function process()
+    {
+        if ($this->order_service->isOrderState(OrderManager::WIRECARD_OS_STARTING)) {
+            $this->order->setCurrentState(\Configuration::get('PS_OS_CANCELED'));
+            $cart_clone = $this->order_service->getNewCartDuplicate();
+            $this->context_service->setCart($cart_clone);
+
+            \Tools::redirect('index.php?controller=order');
+        }
+
+        throw new \Exception('The order is not cancalable');
+    }
 }
