+++ wirecardpaymentgateway/classes/Response/ProcessablePaymentResponse.php
@@ -33,59 +33,16 @@
  * @license GPLv3
  */
 
-namespace WirecardEE\Prestashop\Classes\ResponseProcessing;
-
-use Wirecard\PaymentSdk\Response\FormInteractionResponse;
-use WirecardEE\Prestashop\Helper\Service\ContextService;
+namespace WirecardEE\Prestashop\Classes\Response;
 
 /**
- * Class FormInteractionResponseProcessing
- * @package WirecardEE\Prestashop\Classes\ResponseProcessing
- * @since 2.1.0
+ * Interface ProcessableResponse
+ * @package WirecardEE\Prestashop\Classes\Response
  */
-final class FormInteractionResponseProcessing implements ResponseProcessing
+interface ProcessablePaymentResponse
 {
-    const FORM_TEMPLATE = _PS_MODULE_DIR_ . 'wirecardpaymentgateway' . DIRECTORY_SEPARATOR .
-    'views' . DIRECTORY_SEPARATOR . 'templates' . DIRECTORY_SEPARATOR . 'front' . DIRECTORY_SEPARATOR .
-    'creditcard_submitform.tpl';
-
-    /** @var FormInteractionResponse  */
-    private $response;
-
-    /**
-     * FormInteractionResponseProcessing constructor.
-     *
-     * @param FormInteractionResponse $response
-     */
-    public function __construct($response)
-    {
-        $this->response = $response;
-    }
-
     /**
      * @since 2.1.0
      */
-    public function process()
-    {
-        $context_service = new ContextService(\Context::getContext());
-
-        $context_service->showTemplateWithData(
-            self::FORM_TEMPLATE,
-            $this->getDataFromResponse($this->response)
-        );
-    }
-
-    /**
-     * @param FormInteractionResponse $response
-     * @return array
-     * @since 2.1.0
-     */
-    private function getDataFromResponse($response)
-    {
-        return [
-            'url' => $response->getUrl(),
-            'method' => $response->getMethod(),
-            'form_fields' => $response->getFormFields()
-        ];
-    }
+    public function process();
 }
